moved subsetReducer to the Context
